updated file locations
