--- models/U-Net/pyvenv.cfg
@@ -1,3 +1,0 @@
-home = C:\ProgramData\anaconda3
-include-system-site-packages = false
-version = 3.10.11
